+++ main.tf
@@ -3,17 +3,17 @@
 resource "aws_vpc" "cato-vpc" {
   count      = var.vpc_id == null ? 1 : 0
   cidr_block = var.vpc_range
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-VPC"
-  }
+  })
 }
 
 # Internet Gateway
 resource "aws_internet_gateway" "internet_gateway" {
   count = var.internet_gateway_id == null ? 1 : 0
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-IGW"
-  }
+  })
   vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
 }
 
@@ -23,40 +23,58 @@ data "aws_availability_zones" "available_zones" {
 }
 
 # Subnets
-resource "aws_subnet" "mgmt_subnet" {
+resource "aws_subnet" "mgmt_subnet_primary" {
   vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  cidr_block        = var.subnet_range_mgmt
+  cidr_block        = var.subnet_range_mgmt_primary
   availability_zone = data.aws_availability_zones.available.names[0]
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-MGMT-Subnet"
-  }
+  })
 }
 
-resource "aws_subnet" "wan_subnet" {
+resource "aws_subnet" "mgmt_subnet_secondary" {
   vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  cidr_block        = var.subnet_range_wan
+  cidr_block        = var.subnet_range_mgmt_secondary
+  availability_zone = data.aws_availability_zones.available.names[1]
+  tags = merge(var.tags, {
+    Name = "${var.site_name}-MGMT-Subnet"
+  })
+}
+
+resource "aws_subnet" "wan_subnet_primary" {
+  vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
+  cidr_block        = var.subnet_range_wan_primary
   availability_zone = data.aws_availability_zones.available.names[0]
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-WAN-Subnet"
-  }
+  })
+}
+
+resource "aws_subnet" "wan_subnet_secondary" {
+  vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
+  cidr_block        = var.subnet_range_wan_secondary
+  availability_zone = data.aws_availability_zones.available.names[1]
+  tags = merge(var.tags, {
+    Name = "${var.site_name}-WAN-Subnet"
+  })
 }
 
 resource "aws_subnet" "lan_subnet_primary" {
   vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   cidr_block        = var.subnet_range_lan_primary
   availability_zone = data.aws_availability_zones.available.names[0]
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-LAN-Subnet-Primary"
-  }
+  })
 }
 
 resource "aws_subnet" "lan_subnet_secondary" {
   vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   cidr_block        = var.subnet_range_lan_secondary
-  availability_zone = data.aws_availability_zones.available.names[0]
-  tags = {
+  availability_zone = data.aws_availability_zones.available.names[1]
+  tags = merge(var.tags, {
     Name = "${var.site_name}-LAN-Subnet-Secondary"
-  }
+  })
 }
 
 # Internal and External Security Groups
@@ -90,45 +108,55 @@ resource "aws_security_group" "internal_sg" {
       self             = false
     }
   ]
-  tags = {
-    name = "${var.site_name}-Internal-SG"
-  }
+  tags = merge(var.tags, {
+    name = "${var.site_name}-Cato-Internal-SG"
+  })
 }
 
-resource "aws_security_group" "external_sg_mgmt" {
+resource "aws_security_group" "external_sg" {
   name        = "${var.site_name}-External-SG-MGMT"
   description = "CATO MGMT Security Group - Allow HTTPS In"
   vpc_id      = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  ingress = [
+  ingress     = []
+  egress = [
     {
-      description      = "Allow HTTPS In"
+      description      = "Allow HTTPS Outbound"
       protocol         = "tcp"
       from_port        = 443
       to_port          = 443
-      cidr_blocks      = var.ingress_cidr_blocks
+      cidr_blocks      = ["0.0.0.0/0"]
       ipv6_cidr_blocks = []
       prefix_list_ids  = []
       security_groups  = []
       self             = false
     },
     {
-      description      = "Allow SSH In"
+      description      = "Allow DTLS Outbound"
+      protocol         = "udp"
+      from_port        = 443
+      to_port          = 443
+      cidr_blocks      = ["0.0.0.0/0"]
+      ipv6_cidr_blocks = []
+      prefix_list_ids  = []
+      security_groups  = []
+      self             = false
+    },
+    {
+      description      = "Allow DNS-UDP Outbound"
+      protocol         = "udp"
+      from_port        = 53
+      to_port          = 53
+      cidr_blocks      = ["0.0.0.0/0"]
+      ipv6_cidr_blocks = []
+      prefix_list_ids  = []
+      security_groups  = []
+      self             = false
+    },
+    {
+      description      = "Allow DNS-TCP Outbound"
       protocol         = "tcp"
-      from_port        = 22
-      to_port          = 22
-      cidr_blocks      = var.ingress_cidr_blocks
-      ipv6_cidr_blocks = []
-      prefix_list_ids  = []
-      security_groups  = []
-      self             = false
-    }
-  ]
-  egress = [
-    {
-      description      = "Allow all traffic Outbound"
-      protocol         = -1
-      from_port        = 0
-      to_port          = 0
+      from_port        = 53
+      to_port          = 53
       cidr_blocks      = ["0.0.0.0/0"]
       ipv6_cidr_blocks = []
       prefix_list_ids  = []
@@ -136,72 +164,49 @@ resource "aws_security_group" "external_sg_mgmt" {
       self             = false
     }
   ]
-  tags = {
-    name = "${var.site_name}-External-SG-MGMT"
-  }
-}
-
-resource "aws_security_group" "external_sg_wan" {
-  name        = "${var.site_name}-External-SG-WAN"
-  description = "CATO WAN Security Group - Allow all out, none in"
-  vpc_id      = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  egress = [
-    {
-      description      = "Allow all traffic Outbound"
-      protocol         = -1
-      from_port        = 0
-      to_port          = 0
-      cidr_blocks      = ["0.0.0.0/0"]
-      ipv6_cidr_blocks = []
-      prefix_list_ids  = []
-      security_groups  = []
-      self             = false
-    }
-  ]
-  tags = {
-    name = "${var.site_name}-External-SG-WAN"
-  }
+  tags = merge(var.tags, {
+    name = "${var.site_name}-Cato-External-SG"
+  })
 }
 
 # vSocket Network Interfaces
 resource "aws_network_interface" "mgmteni_primary" {
   source_dest_check = "true"
-  subnet_id         = aws_subnet.mgmt_subnet.id
+  subnet_id         = aws_subnet.mgmt_subnet_primary.id
   private_ips       = [var.mgmt_eni_primary_ip]
-  security_groups   = [aws_security_group.external_sg_mgmt.id]
-  tags = {
-    Name = "${var.site_name}-MGMT-INT-Primary"
-  }
+  security_groups   = [aws_security_group.external_sg.id]
+  tags = merge(var.tags, {
+  Name = "${var.site_name}-MGMT-INT-Primary" })
 }
 
 resource "aws_network_interface" "mgmteni_secondary" {
   source_dest_check = "true"
-  subnet_id         = aws_subnet.mgmt_subnet.id
+  subnet_id         = aws_subnet.mgmt_subnet_secondary.id
   private_ips       = [var.mgmt_eni_secondary_ip]
-  security_groups   = [aws_security_group.external_sg_mgmt.id]
-  tags = {
+  security_groups   = [aws_security_group.external_sg.id]
+  tags = merge(var.tags, {
     Name = "${var.site_name}-MGMT-INT-Secondary"
-  }
+  })
 }
 
 resource "aws_network_interface" "waneni_primary" {
   source_dest_check = "true"
-  subnet_id         = aws_subnet.wan_subnet.id
+  subnet_id         = aws_subnet.wan_subnet_primary.id
   private_ips       = [var.wan_eni_primary_ip]
-  security_groups   = [aws_security_group.external_sg_wan.id]
-  tags = {
+  security_groups   = [aws_security_group.external_sg.id]
+  tags = merge(var.tags, {
     Name = "${var.site_name}-WAN-INT-Primary"
-  }
+  })
 }
 
 resource "aws_network_interface" "waneni_secondary" {
   source_dest_check = "true"
-  subnet_id         = aws_subnet.wan_subnet.id
+  subnet_id         = aws_subnet.wan_subnet_secondary.id
   private_ips       = [var.wan_eni_secondary_ip]
-  security_groups   = [aws_security_group.external_sg_wan.id]
-  tags = {
+  security_groups   = [aws_security_group.external_sg.id]
+  tags = merge(var.tags, {
     Name = "${var.site_name}-WAN-INT-Secondary"
-  }
+  })
 }
 
 
@@ -210,9 +215,9 @@ resource "aws_network_interface" "laneni_primary" {
   subnet_id         = aws_subnet.lan_subnet_primary.id
   private_ips       = [var.lan_eni_primary_ip]
   security_groups   = [aws_security_group.internal_sg.id]
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-LAN-INT-Primary"
-  }
+  })
 }
 
 resource "aws_network_interface" "laneni_secondary" {
@@ -220,34 +225,34 @@ resource "aws_network_interface" "laneni_secondary" {
   subnet_id         = aws_subnet.lan_subnet_secondary.id
   private_ips       = [var.lan_eni_secondary_ip]
   security_groups   = [aws_security_group.internal_sg.id]
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-LAN-INT-Secondary"
-  }
+  })
 }
 
 # Elastic IP Addresses
 resource "aws_eip" "waneip_primary" {
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-WAN-EIP-Primary"
-  }
+  })
 }
 
 resource "aws_eip" "waneip_secondary" {
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-WAN-EIP-Secondary"
-  }
+  })
 }
 
 resource "aws_eip" "mgmteip_primary" {
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-MGMT-EIP-Primary"
-  }
+  })
 }
 
 resource "aws_eip" "mgmteip_secondary" {
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-MGMT-EIP-Secondary"
-  }
+  })
 }
 
 # Elastic IP Addresses Association - Required to properly destroy 
@@ -274,34 +279,21 @@ resource "aws_eip_association" "mgmteip_assoc_secondary" {
 # Routing Tables
 resource "aws_route_table" "wanrt" {
   vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-WAN-RT"
-  }
-}
-
-resource "aws_route_table" "mgmtrt" {
-  vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  tags = {
-    Name = "${var.site_name}-MGMT-RT"
-  }
+  })
 }
 
 resource "aws_route_table" "lanrt" {
   vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
-  tags = {
+  tags = merge(var.tags, {
     Name = "${var.site_name}-LAN-RT"
-  }
+  })
 }
 
 # Routes
 resource "aws_route" "wan_route" {
   route_table_id         = aws_route_table.wanrt.id
-  destination_cidr_block = "0.0.0.0/0"
-  gateway_id             = var.internet_gateway_id == null ? aws_internet_gateway.internet_gateway[0].id : var.internet_gateway_id
-}
-
-resource "aws_route" "mgmt_route" {
-  route_table_id         = aws_route_table.mgmtrt.id
   destination_cidr_block = "0.0.0.0/0"
   gateway_id             = var.internet_gateway_id == null ? aws_internet_gateway.internet_gateway[0].id : var.internet_gateway_id
 }
@@ -313,13 +305,23 @@ resource "aws_route" "lan_route" {
 }
 
 # Route Table Associations
-resource "aws_route_table_association" "mgmt_subnet_route_table_association" {
-  subnet_id      = aws_subnet.mgmt_subnet.id
-  route_table_id = aws_route_table.mgmtrt.id
+resource "aws_route_table_association" "mgmt_subnet_primary_route_table_association" {
+  subnet_id      = aws_subnet.mgmt_subnet_primary.id
+  route_table_id = aws_route_table.wanrt.id
 }
 
-resource "aws_route_table_association" "wan_subnet_route_table_association" {
-  subnet_id      = aws_subnet.wan_subnet.id
+resource "aws_route_table_association" "mgmt_subnet_secondary_route_table_association" {
+  subnet_id      = aws_subnet.mgmt_subnet_secondary.id
+  route_table_id = aws_route_table.wanrt.id
+}
+
+resource "aws_route_table_association" "wan_subnet_primary_route_table_association" {
+  subnet_id      = aws_subnet.wan_subnet_primary.id
+  route_table_id = aws_route_table.wanrt.id
+}
+
+resource "aws_route_table_association" "wan_subnet_secondary_route_table_association" {
+  subnet_id      = aws_subnet.wan_subnet_secondary.id
   route_table_id = aws_route_table.wanrt.id
 }
 
@@ -342,7 +344,7 @@ resource "cato_socket_site" "aws-site" {
   description     = var.site_description
   name            = var.site_name
   native_range = {
-    native_network_range = var.subnet_range_lan_primary
+    native_network_range = var.native_network_range
     local_ip             = var.lan_eni_primary_ip
   }
   site_location = var.site_location
@@ -379,6 +381,9 @@ resource "aws_iam_role" "cato_ha_role" {
       }
     }]
   })
+  tags = merge(var.tags, {
+    Name = "Cato-HA-Role-${local.sanitized_name}"
+  })
 }
 
 resource "aws_iam_policy" "cato_ha_policy" {
@@ -397,6 +402,9 @@ resource "aws_iam_policy" "cato_ha_policy" {
       },
     ]
   })
+  tags = merge(var.tags, {
+    Name = "Cato-HA-Role-Policy-${local.sanitized_name}"
+  })
 }
 
 resource "aws_iam_role_policy_attachment" "cato_ha_attach" {
@@ -407,6 +415,9 @@ resource "aws_iam_role_policy_attachment" "cato_ha_attach" {
 resource "aws_iam_instance_profile" "cato_ha_instance_profile" {
   name = "Cato-HA-Role-${local.sanitized_name}"
   role = aws_iam_role.cato_ha_role.name
+  tags = merge(var.tags, {
+    Name = "Cato-HA-Role-${local.sanitized_name}"
+  })
 }
 
 ## Lookup data from region and VPC