+++ variables.tf
@@ -30,6 +30,14 @@ variable "vpc_range" {
   default     = null
 }
 
+variable "native_network_range" {
+  type        = string
+  description = <<EOT
+  	Choose a unique range for your new vsocket site that does not conflict with the rest of your Wide Area Network.
+    The accepted input format is Standard CIDR Notation, e.g. X.X.X.X/X
+	EOT
+}
+
 variable "internet_gateway_id" {
   description = "Specify an Internet Gateway ID to use. If not specified, a new Internet Gateway will be created."
   type        = string
@@ -62,6 +70,7 @@ variable "site_type" {
 }
 
 variable "site_location" {
+  description = "The Site Location Data"
   type = object({
     city         = string
     country_code = string
@@ -137,7 +146,8 @@ variable "lan_eni_secondary_ip" {
   type        = string
 }
 
-variable "subnet_range_mgmt" {
+
+variable "subnet_range_mgmt_primary" {
   type        = string
   description = <<EOT
     Choose a range within the VPC to use as the Management subnet. This subnet will be used initially to access the public internet and register your vSocket to the Cato Cloud.
@@ -145,8 +155,23 @@ variable "subnet_range_mgmt" {
     The accepted input format is Standard CIDR Notation, e.g. X.X.X.X/X
 	EOT
 }
-
-variable "subnet_range_wan" {
+variable "subnet_range_mgmt_secondary" {
+  type        = string
+  description = <<EOT
+    Choose a range within the VPC to use as the Management subnet. This subnet will be used initially to access the public internet and register your vSocket to the Cato Cloud.
+    The minimum subnet length to support High Availability is /28.
+    The accepted input format is Standard CIDR Notation, e.g. X.X.X.X/X
+	EOT
+}
+variable "subnet_range_wan_primary" {
+  type        = string
+  description = <<EOT
+    Choose a range within the VPC to use as the Public/WAN subnet. This subnet will be used to access the public internet and securely tunnel to the Cato Cloud.
+    The minimum subnet length to support High Availability is /28.
+    The accepted input format is Standard CIDR Notation, e.g. X.X.X.X/X
+	EOT
+}
+variable "subnet_range_wan_secondary" {
   type        = string
   description = <<EOT
     Choose a range within the VPC to use as the Public/WAN subnet. This subnet will be used to access the public internet and securely tunnel to the Cato Cloud.