+++ changelog.md
@@ -27,3 +27,32 @@
 
 ### Features
 - Fixed typp in readme
+
+## 0.0.10 (2025-05-19)
+
+### Features
+ - Adjusted Security Groups to be in-line with ProServ recommendations.
+ - Removed Mgmt Security Group (Unneeded)
+ - Enhanced Tagging to Tagable resources. 
+ - Removed Mgmt RouteTable (Unneeded)
+ - Simplified Outputs 
+
+ - Prepared Module for Use in TGW HA Module 
+   - Added subnets for secondary mgmt (Separate AZ) 
+   - Added subnets for secondary wan (Separate AZ)
+   - Adjusted Route Table Association to accomodate new Subnets
+   - Changed Native Network Range to use var "native_network_range" 
+   - Added additional outputs 
+     - mgmt_subnet_primary_id 
+     - mgmt_subnet_secondary_id
+     - wan_subnet_primary_id
+     - wan_subnet_secondary_id
+    - Added additional required variables 
+      - native_network_range
+      - subnet_range_mgmt_primary
+      - subnet_range_mgmt_secondary
+      - subnet_range_wan_primary
+      - subnet_range_wan_secondary
+ - Updated ChangeLog
+ - Updated Readme 
+